modify

+++ src/data_cell/sparse_term_datacell.cpp
@@ -34,11 +34,12 @@ SparseTermDataCell::Query(float* global_dists, const SparseTermComputerPtr& comp
         if (term >= term_ids_.size()) {
             continue;
         }
-        computer->ScanForAccumulate(it,
-                                    term_ids_[term].data(),
-                                    term_datas_[term].data(),
-                                    term_sizes_[term] * computer->term_prune_ratio_,
-                                    global_dists);
+        computer->ScanForAccumulate(
+            it,
+            term_ids_[term].data(),
+            term_datas_[term].data(),
+            static_cast<float>(term_sizes_[term]) * computer->term_prune_ratio_,
+            global_dists);
     }
     computer->ResetTerm();
 }
@@ -72,7 +73,8 @@ SparseTermDataCell::InsertHeap(float* dists,
         uint32_t i = 0;
         if constexpr (mode == InnerSearchMode::KNN_SEARCH) {
             if (heap.size() < n_candidate) {
-                for (; i < term_sizes_[term] * computer->term_prune_ratio_; i++) {
+                for (; i < static_cast<float>(term_sizes_[term]) * computer->term_prune_ratio_;
+                     i++) {
                     id = term_ids_[term][i];
 
                     heap.emplace(dists[id], id + offset_id);
@@ -86,7 +88,7 @@ SparseTermDataCell::InsertHeap(float* dists,
             }
         }
 
-        for (; i < term_sizes_[term] * computer->term_prune_ratio_; i++) {
+        for (; i < static_cast<float>(term_sizes_[term]) * computer->term_prune_ratio_; i++) {
             id = term_ids_[term][i];
 
             if (dists[id] >= cur_heap_top) [[likely]] {
